+++ tools/maint/run_python.sh
@@ -30,4 +30,5 @@ if [ -z "$_EM_PY" ]; then
   exit 1
 fi
 
+export _EM_LAUNCHER=1
 exec "$_EM_PY" -E "$0.py" "$@"